refactor: move HarnessCoverage from coverage.py to types.py
types.py is the canonical location for shared TypedDicts but was
importing HarnessCoverage from coverage.py, creating a layer violation.
Move the definition to types.py and update import sites.

Closes #838
